requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
h5py==3.10.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy>=1.24.3
Pillow==10.2.0
tensorflow>=2.15.0
Werkzeug==3.0.1
